--- a/openquake/baselib/workerpool.py
+++ b/openquake/baselib/workerpool.py
@@ -31,21 +31,6 @@ except ImportError:
         "Do nothing"
 
 
-def _streamer(ctrl_port):
-    # streamer for zmq workers running on the master node
-    task_input_url = 'tcp://0.0.0.0:%d' % (ctrl_port + 2)
-    task_output_url = 'tcp://%s:%s' % (config.dbserver.listen, ctrl_port + 1)
-    if (general.socket_ready(('0.0.0.0', ctrl_port + 1)) or
-            general.socket_ready(('0.0.0.0', ctrl_port + 2))):
-        return  # already started
-    sock_in = z.bind(task_input_url, z.zmq.PULL)
-    sock_out = z.bind(task_output_url, z.zmq.PUSH)
-    try:
-        z.zmq.proxy(sock_in, sock_out)
-    except (KeyboardInterrupt, z.zmq.ContextTerminated):
-        pass  # killed cleanly by SIGINT/SIGTERM
-
-
 def ssh_args(zworkers):
     """
     :yields: triples (hostIP, num_cores, [ssh remote python command])
@@ -227,7 +212,9 @@ class WorkerPool(object):
 
 
 def workerpool(worker_url='tcp://0.0.0.0:1909', *, num_workers: int = -1):
-    # start a workerpool without a streamer
+    """
+    Start a workerpool on the given URL with the given number of workers.
+    """
     WorkerPool(worker_url, num_workers).start()