By removing the task streamer and using a regular REQ/REP socket talking with a normal process pool we can use all of the cluster (and not 60% of it as it was happening for the AUS model). Also, we can save a lot of memory in the master node.
Here is a 21% speedup for the Alaska calculation:
```
# before
| calc_49738, maxmem=14.4 GB | time_sec | memory_mb | counts    |
|----------------------------+----------+-----------+-----------|
| total classical            | 14_460   | 305.0     | 480       |
| composing pnes             | 6_726    | 0.0       | 9_500     |
| collapsing contexts        | 3_004    | 232.8     | 9_500     |
| nonplanar contexts         | 1_907    | 0.0       | 46_638    |
| planar contexts            | 1_845    | 0.0       | 287_884   |
| iter_ruptures              | 500.6    | 0.92969   | 9_569     |
| ClassicalCalculator.run    | 436.9    | 722.4     | 1         |

# after
| calc_49737, maxmem=14.5 GB | time_sec | memory_mb | counts    |
|----------------------------+----------+-----------+-----------|
| total classical            | 14_593   | 306.3     | 480       |
| composing pnes             | 6_773    | 0.0       | 9_500     |
| collapsing contexts        | 3_060    | 232.8     | 9_500     |
| nonplanar contexts         | 1_922    | 0.0       | 46_638    |
| planar contexts            | 1_872    | 0.0       | 287_884   |
| iter_ruptures              | 508.5    | 0.62891   | 9_569     |
| ClassicalCalculator.run    | 344.4    | 831.0     | 1         |
```